Fix pointer-to-integer truncation in the gsg script round-trip
grid_ui_event_get_script passes the address of a stack buffer to the Lua
"gsg" helper as an integer, which writes the script source back into it.
Both ends cast through uint32_t, which truncates the pointer on 64-bit
builds (the host unit-test build warns with -Wpointer-to-int-cast). On the
32-bit target pointers fit, so this was silent there but is a latent bug
off-target.

Route the pointer through uintptr_t on both sides and format it as
unsigned long long (%llu), which matches Lua's 64-bit lua_Integer. Warning-
free and correct on both 32- and 64-bit; unchanged behavior on the target.

+++ common/src/c/grid_lua_api.c
@@ -517,7 +517,7 @@ int l_grid_cat(lua_State* L) {
 
     if (lua_type(L, 1) == LUA_TNUMBER && lua_type(L, 2) == LUA_TSTRING) {
 
-      uint32_t pointer = (uint32_t)lua_tointeger(L, 1);
+      uintptr_t pointer = (uintptr_t)lua_tointeger(L, 1);
 
       char* string = (char*)pointer;
